+++ CHANGES.md
@@ -20,6 +20,7 @@ Thread safe, reentrant, modifiable.
 - TODO updates re: Humanizing Data
 - Added migration of statefiles to new naming convention
 - Added explicit configuration testing option
+- enabled handling multiple site config files in one invocation
 
 # Release 1.3.7 (23-Mar-26): More MRTG UX Tweaks
 - Added PayPal donation link to MRTG site